Remove WixComponent
