+++ src/common.d.ts
@@ -9,10 +9,6 @@ export interface InjectedFocusableProps {
 
 export type IconElement = React.ReactElement<any>;
 
-export type WixComponentClickOutsideEventHandler = (
-  e: TouchEvent | MouseEvent,
-) => void;
-
 export type OmitPolyfill<T, K extends keyof any> = Pick<T, Exclude<keyof T, K>>;
 
 export interface FocusOptionsPolyfill {
